Originating revision: r404 | nedim_fresko | 2006-11-03 15:07:10 -0800 (Fri, 03 Nov 2006) | 18 lines
Package up a stand-alone native library for JUMP.
                                                                                
Native applications and other processes (like JSR drivers) can link
against this library to communicate with other JUMP processes. The
target is:
                                                                                
<cdc output>/jump/lib/libjumpmesg.so
                                                                                
(or libjumpmesg_g.so for a debug build).
                                                                                
Also made sure cvmc gets the objects in this library.
                                                                                
The next step is to make sure the messaging library is used for the
mtask server and cvmc.
                                                                                
Reviewed by Chris.